# app/swagger/swagger.yml
swagger: "2.0"
info:
  title: "Module task API"
  description: "API documentation for the module task."
  version: "1.0"
paths:
...
  /api/user:
    post:
      summary: "Create a new user"
      tags: 
        - "User"
      consumes:
        - "application/json"
      parameters:
        - in: body
          name: user 
          schema:
            type: object
            properties:
              fullName:
                type: string
                description: "The full name of the user."
                example: "Your full name"
              email:
                type: string
                description: "The email of the user."
                example: "Your email"
              password:
                type: string
                description: "The password of the user."
                example: "Your password"
              role:
                type: string
                description: "The role of the user."
                example: "Your role"
              username:
                type: string
                description: "The username of the user."
                example: "Your username"
              age:
                type: integer
                description: "The age of the user."
                example: 18
              avatarImg:
                type: string
                description: "The avatar image of the user."
                example: "Your avatar image link"
      required: true
      responses:
        201:
          description: "User created successfully."
          schema:
            type: object
            properties:
              _id:
                type: string
                description: "The ID of the user."
              fullName:
                type: string
                description: "The full name of the user."
              email:
                type: string
                description: "The email of the user."
              role:
                type: string
                description: "The role of the user."
              username:
                type: string
                description: "The username of the user."
              age:
                type: integer
                description: "The age of the user."
              avatarImg:
                type: string
                description: "The avatar image of the user."
              exp:
                type: integer
                description: "The experience of the user"
              level:
                type: integer
                description: "The level of the user"
              challengesCompleted:
                type: integer
                description: "The number of completed challenges by the user"
    get:
      summary: "Get all users"
      tags: 
        - "User"
      consumes:
        - "application/json"
      responses:
        200:
          description: "All users returned successfully"
          schema:
            type: object
            properties:
              _id:
                type: string
                description: "The ID of the user."
              fullName:
                type: string
                description: "The full name of the user."
              email:
                type: string
                description: "The email of the user."
              role:
                type: string
                description: "The role of the user."
              username:
                type: string
                description: "The username of the user."
              age:
                type: integer
                description: "The age of the user."
              avatarImg:
                type: string
                description: "The avatar image of the user."
              exp:
                type: integer
                description: "The experience of the user"
              level:
                type: integer
                description: "The level of the user"
              challengesCompleted:
                type: integer
                description: "The number of completed challenges by the user"
  /api/user/{userId}:
    get:
      summary: "Get user by userId"
      tags: 
        - "User"
      consumes:
        - "application/json"
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: String ID of the user to get
      responses:
        200:
          description: "User returned successfully"
          schema:
            type: object
            properties:
              _id:
                type: string
                description: "The ID of the user."
              fullName:
                type: string
                description: "The full name of the user."
              email:
                type: string
                description: "The email of the user."
              role:
                type: string
                description: "The role of the user."
              username:
                type: string
                description: "The username of the user."
              age:
                type: integer
                description: "The age of the user."
              avatarImg:
                type: string
                description: "The avatar image of the user."
              exp:
                type: integer
                description: "The experience of the user"
              level:
                type: integer
                description: "The level of the user"
              challengesCompleted:
                type: integer
                description: "The number of completed challenges by the user"
    delete:
      summary: "Delete user by userId"
      tags: 
        - "User"
      consumes:
        - "application/json"
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: String ID of the user to delete
      responses:
        200:
          description: "User deleted successfully"
          schema:
            type: object
            properties:
              message:
                type: string
                description: "Delete success message"
    put: 
      summary: "Change user by userId"
      tags: 
        - "User"
      consumes:
        - "application/json"
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: String ID of the user to change
        - in: body
          name: user
          schema:
            type: object
            properties:
              fullName:
                type: string
                description: "The full name of the user."
                example: "Your full name"
              email:
                type: string
                description: "The email of the user."
                example: "Your email"
              password:
                type: string
                description: "The password of the user."
                example: "Your password"
              role:
                type: string
                description: "The role of the user."
                example: "Your role"
              username:
                type: string
                description: "The username of the user."
                example: "Your username"
              age:
                type: integer
                description: "The age of the user."
                example: 18
              avatarImg:
                type: string
                description: "The avatar image of the user."
                example: "Your avatar image link"
              exp:
                type: integer
                description: "The experience of the user."
                example: 0
              level:
                type: integer
                description: "The level of the user."
                example: 0
              challengesCompleted:
                type: string
                description: "The number of completed challenges by the user."
                example: 0
      responses:
        200:
          description: "User changed successfully"
          schema:
            type: object
            properties:
              _id:
                type: string
                description: "The ID of the user."
              fullName:
                type: string
                description: "The full name of the user."
              email:
                type: string
                description: "The email of the user."
              role:
                type: string
                description: "The role of the user."
              username:
                type: string
                description: "The username of the user."
              age:
                type: integer
                description: "The age of the user."
              avatarImg:
                type: string
                description: "The avatar image of the user."
              exp:
                type: integer
                description: "The experience of the user"
              level:
                type: integer
                description: "The level of the user"
              challengesCompleted:
                type: integer
                description: "The number of completed challenges by the user"
  /api/user/completeChallenge/{userId}:
    put: 
      summary: "User complete challenge and calling this route"
      tags: 
        - "User"
      consumes:
        - "application/json"
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: String ID of the user to change
        - in: body
          name: user
          schema:
            type: object
            properties:
              challengeId:
                type: string
                description: "String ID of the completed challenge by user"
                example: "Completed challenge id"
      responses:
        200:
          description: "User changed successfully"
          schema:
            type: object
            properties:
              _id:
                type: string
                description: "The ID of the user."
              fullName:
                type: string
                description: "The full name of the user."
              email:
                type: string
                description: "The email of the user."
              role:
                type: string
                description: "The role of the user."
              username:
                type: string
                description: "The username of the user."
              age:
                type: integer
                description: "The age of the user."
              avatarImg:
                type: string
                description: "The avatar image of the user."
              exp:
                type: integer
                description: "The experience of the user"
              level:
                type: integer
                description: "The level of the user"
              challengesCompleted:
                type: integer
                description: "The number of completed challenges by the user"